ignoring user left message
