+++ visual_tests/src/test/java/base/BaseTests.java
@@ -3,7 +3,9 @@ package base;
 import com.applitools.eyes.BatchInfo;
 import com.applitools.eyes.RectangleSize;
 import com.applitools.eyes.selenium.Eyes;
+import com.applitools.eyes.selenium.fluent.Target;
 import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.By;
 import pages.HomePage;
 
 import org.junit.BeforeClass;
@@ -23,34 +25,54 @@ public class BaseTests {
   protected static WebDriver driver;
   protected static Eyes eyes;
   protected static HomePage homePage;
-  protected static ChromeOptions options = new ChromeOptions();
-  private static String _runWhere;
+  private static String runWhere;
+  private static ChromeOptions chromeOptions;
+  protected static String testStartPage;
 
   @BeforeClass
-  public static void setUp() throws MalformedURLException, InterruptedException{
-  
-    _runWhere = System.getenv("RUNWHERE");
-    ChromeOptions ChromeOptions = new ChromeOptions();
+  public static void setSuite(){
+    testStartPage = System.getenv().get("TEST_START_PAGE");
+
+    // obtain the batch name and ID from the environment variables
+    String batchName = System.getenv("APPLITOOLS_BATCH_NAME");
+    String batchId   = System.getenv("APPLITOOLS_BATCH_ID");
+
+    // set the batch
+    BatchInfo batchInfo = new BatchInfo(batchName);
+    batchInfo.setId(batchId);
+    eyes.setBatch(batchInfo);
+  }
+
+  @BeforeClass
+  public static void setUp() throws MalformedURLException{
+
+    chromeOptions = new ChromeOptions();
     WebDriverManager.chromedriver().setup();
 
-    if (_runWhere.equals("local")) {
-      // Standard local visual test call
-      driver = new ChromeDriver();
-    } 
-    else if (_runWhere.equals("pipeline")) {
-      // build server headless chrome CI/CD example
-      ChromeOptions.addArguments("--headless", "--no-sandbox");
-      driver = new ChromeDriver(ChromeOptions);
-    }
-    else if (_runWhere.equals("container")) {
-      // selenium hub remote settings (container based CI/CD)
-      String Selenium = "http://selenium_hub:4444/wd/hub";
-      driver = new RemoteWebDriver(new URL(Selenium), ChromeOptions);
-    };
+    getEnvironment();
       
     // For use with Applitools
     eyes = new Eyes();
     homePage = new HomePage(driver);
+  }
+
+  private static void getEnvironment() throws MalformedURLException {
+    runWhere = System.getenv("RUNWHERE");
+
+    if (runWhere.equals("local")) {
+      // Standard local visual test call
+      driver = new ChromeDriver();
+    }
+    else if (runWhere.equals("pipeline")) {
+      // build server headless chrome CI/CD example
+      chromeOptions.addArguments("--headless", "--no-sandbox");
+      driver = new ChromeDriver(chromeOptions);
+    }
+    else if (runWhere.equals("container")) {
+      // selenium hub remote settings (container based CI/CD)
+      String Selenium = "http://selenium_hub:4444/wd/hub";
+      driver = new RemoteWebDriver(new URL(Selenium), chromeOptions);
+    };
   }
 
   @AfterClass
@@ -65,7 +87,7 @@ public class BaseTests {
         "ChattyBot", 
         Thread.currentThread().getStackTrace()[2].getMethodName(),
         new RectangleSize(1200, 774));
-    eyes.checkWindow();
+    eyes.check(Target.window().ignore(By.xpath("//ul/li/p[text()='angie left!']")));
     eyes.close();
   }
